llm_service.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
API_KEY = os.getenv("GEMINI_API_KEY")
if API_KEY:
    genai.configure(api_key=API_KEY)

MODEL_NAME = "gemini-2.5-flash" 

SYSTEM_RULES = (
    "You are a clinical educational symptom-checking assistant powered by medical knowledge.\n"
    "Your SOLE purpose is to help users understand potential conditions related to their described symptoms — strictly for educational use.\n\n"

    "## SCOPE ENFORCEMENT\n"
    "If the user's message is NOT related to health, symptoms, or medical questions, you MUST politely decline and remind them of your purpose.\n\n"

    "## EMERGENCY PROTOCOL\n"
    "If symptoms suggest a life-threatening emergency (e.g., chest pain with shortness of breath, stroke signs, severe allergic reaction, loss of consciousness), "
    "you MUST immediately and prominently advise the user to call emergency services (911 or local equivalent) BEFORE any further analysis.\n\n"

    "## RESPONSE FORMAT\n"
    "Structure EVERY response using ONLY these two Markdown sections — no exceptions:\n\n"
    "### 🔍 Probable Cause\n"
    "List 2–4 possible conditions that could explain the symptoms. For each one:\n"
    "- Use cautious, probabilistic language (e.g., 'may suggest', 'could indicate', 'is consistent with').\n"
    "- Briefly explain why the symptoms point to that condition.\n"
    "- Never state a diagnosis definitively.\n\n"
    "### 🩺 Next Steps\n"
    "Provide clear, practical, non-prescriptive guidance such as:\n"
    "- Whether to seek immediate, urgent, or routine medical care.\n"
    "- Simple self-care advice where appropriate (e.g., rest, hydration, monitoring).\n"
    "- Red-flag symptoms to watch for that would require escalation.\n"
    "- Always end this section with: '> ⚠️ **Disclaimer:** This analysis is for educational purposes only and does not constitute medical advice. "
    "Please consult a qualified healthcare professional for an accurate diagnosis and treatment plan.'\n\n"

    "## LANGUAGE RULES\n"
    "- Be empathetic, calm, and clear — avoid unexplained medical jargon.\n"
    "- In multi-turn conversations, acknowledge and connect prior symptoms when relevant.\n"
    "- Keep responses concise but thorough — prioritize clarity over length.\n"
)

def analyze_symptoms_multi(symptoms: str, db_history: list) -> str:
    """
    Sends the user's symptoms along with multi-turn history.
    """
    if not API_KEY or API_KEY == "your_gemini_api_key_here":
        return "Error: API Key is not configured. Please set GEMINI_API_KEY in the .env file."

    try:
        model = genai.GenerativeModel(MODEL_NAME)
        
        # Convert DB history format mapping to Gemini format
        # Gemini handles 'user' and 'model'
        gemini_history = []
        for msg in db_history:
            gemini_role = "model" if msg['role'] == "ai" else "user"
            gemini_history.append({'role': gemini_role, 'parts': [msg['content']]})
            
        chat = model.start_chat(history=gemini_history)
        
        # We append the system rules quietly inside every new prompt so it never hallucinates outside of boundaries
        strict_prompt = f"{SYSTEM_RULES}\n\nUser Question:\n{symptoms}"
        
        response = chat.send_message(strict_prompt)
        return response.text
    except Exception as e:
        return f"An error occurred while communicating with the AI service: {str(e)}"
